@@ -17,16 +17,16 @@
 		<!-- 编译UTF-8 -->
 		<maven.compiler.encoding>UTF-8</maven.compiler.encoding>
 		<!-- java jdk 版本声明 可变更 根据自己配置去匹配 -->
-		<java.version>1.8</java.version>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<java.version>21</java.version>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
 		<maven.deploy.skip>true</maven.deploy.skip>
 		<maven.install.skip>true</maven.install.skip>
 		<mavenCompilerPlugin.version>3.14.0</mavenCompilerPlugin.version>
 		<mavenSurefirePlugin.version>3.5.2</mavenSurefirePlugin.version>
 
-		<log4j.version>2.17.2</log4j.version>
-		<testng.version>7.5</testng.version>
+		<log4j.version>2.25.2</log4j.version>
+		<testng.version>7.11.0</testng.version>
 
 		<poi.version>5.4.1</poi.version>
 	</properties>
